refactor: centralize send balance state
